test(analytics): split Tier 2 test file by domain + shared helpers (M5, H3)
The original ``test_posthog_tier2_events.py`` reached ~485 lines / 17
tests across 4 unrelated domains. Each new addition pushed it further
past the project's readable-file threshold, and the per-test boilerplate
(``_patch_capture`` / ``_called_events`` / inline ``next(...)``) was
duplicated 17 times.

Split into per-domain files mirroring the event namespaces:

- ``test_posthog_tier2_teams.py`` (335 lines) — team:* and api_token:*
- ``test_posthog_tier2_billing.py`` (52 lines) — billing:trial_expired
- ``test_posthog_tier2_documents.py`` (50 lines) — document:access_denied
- ``test_posthog_tier2_service.py`` (82 lines) — search:performed + the
  service-level guards (anonymous distinct_id, disabled PostHog)

Centralised the boilerplate in ``posthog_helpers.py``:

- ``patch_capture(mocker)`` — replaces the per-file ``_patch_capture``
- ``called_events(mock)`` — list of fired event names
- ``find_call(mock, event)`` — the first call matching ``event``,
  replacing 11 inline ``next(c for c in ... if c.args[1] == ...)`` blocks
- ``assert_workspace_attribution(mock, event, workspace_key)`` — H3
  helper that enforces the Tier 2 convention (distinct_id == workspace
  key AND ``groups={"workspace": workspace_key}``) in one assert with a
  good failure message. Applied to representative tests in each domain
  (member_invited, custom_domain_added, role_changed) so the helper is
  exercised; remaining tests can adopt incrementally.

The old single file is removed. Largest new file (teams, 335 lines)
sits below the 396-line size the auditor flagged.

Tests: all 17 pass across the 4 new files.
